fix C4244 warning

+++ tests/cpp-tests/Classes/BillBoardTest/BillBoardTest.cpp
@@ -137,7 +137,7 @@ BillBoardTest::BillBoardTest()
         auto billboard = BillBoard::create(imgs[(unsigned int)(CCRANDOM_0_1() * 1 + 0.5)]);
         billboard->setScale(0.5f);
         billboard->setPosition3D(Vec3(0.0f, 0.0f,  CCRANDOM_MINUS1_1() * 150.0f));
-        billboard->setOpacity(CCRANDOM_0_1() * 128 + 128);
+        billboard->setOpacity(static_cast<uint8_t>(CCRANDOM_0_1() * 128 + 128));
         _billboards.push_back(billboard);
         layer->addChild(billboard);
         _layerBillBoard->addChild(layer);
@@ -238,10 +238,10 @@ void BillBoardTest::addNewBillBoardWithCoords(Vec3 p)
     std::string imgs[3] = {"Images/Icon.png", "Images/r2.png"};
     for (unsigned int i = 0; i < 10; ++i)
     {
-        auto billboard = BillBoard::create(imgs[(unsigned int)(CCRANDOM_0_1() * 1 + 0.5)]);
+        auto billboard = BillBoard::create(imgs[(unsigned int)(CCRANDOM_0_1() * 1 + 0.5f)]);
         billboard->setScale(0.5f);
         billboard->setPosition3D(Vec3(p.x, p.y, -150.0f + 30 * i));
-        billboard->setOpacity(CCRANDOM_0_1() * 128 + 128);
+        billboard->setOpacity(static_cast<uint8_t>(CCRANDOM_0_1() * 128 + 128));
         
         _layerBillBoard->addChild(billboard);
         _billboards.push_back(billboard);
@@ -269,7 +269,7 @@ void BillBoardTest::addNewAniBillBoardWithCoords(Vec3 p)
 
         auto action = Animate::create(animation);
         billboardAni->runAction(RepeatForever::create(action));
-        billboardAni->setOpacity(CCRANDOM_0_1() * 128 + 128);
+        billboardAni->setOpacity(static_cast<uint8_t>(CCRANDOM_0_1() * 128 + 128));
         _billboards.push_back(billboardAni);
     }
 }
